+++ modules/aws-ec2-instance/variables.tf
@@ -53,10 +53,10 @@ variable "enable_volume_encryption" {
   default     = true
 }
 
-variable "additional_tags" {
+variable "tags" {
   type        = map(string)
-  description = "Additional tags to be attached to the resources created"
-  default     = { Author : "Tamr", Name : "Tamr VM" }
+  description = "A map of tags to add to all resources."
+  default     = {}
 }
 
 variable "bootstrap_scripts" {